Remove Duplicate test_clock.py File and sample test_rpe.py file
