+++ tests/test_clock.py
@@ -1,12 +1,12 @@
 #
 #  Copyright (C) 2024 RapidSilicon
 #  Authorized use only
-
-from submodule.clock import Clock, Source, ClockOutput, Clock_SubModule, Clock_State
-from submodule.rs_device_resources import RsDeviceResources
-from submodule.rs_message import RsMessageType
-from unittest.mock import Mock
+#
 import pytest
+from unittest.mock import Mock
+from submodule.clock import Clock, Source, ClockOutput, Clock_SubModule, Clock_State
+from submodule.rs_device_resources import RsDeviceResources, ClockNotFoundException, ClockMaxCountReachedException, ClockDescriptionPortValidationException
+from submodule.rs_message import RsMessageType
 
 def test_clock_initialization():
     clock = Clock()
@@ -75,3 +75,94 @@ def test_get_clocking_resources():
 
     expected_result = (16, 2, total_clock_used, total_pll_used)
     assert clock_submodule.get_resources() == expected_result
+
+def test_add_clock_success():
+    mock_resources = Mock(spec=RsDeviceResources)
+    mock_resources.get_num_Clocks.return_value = 5
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    data = {"description": "Clock A", "port": "PORT_A", "enable": True}
+    clock = clock_submodule.add(data)
+
+    assert clock.description == "Clock A"
+    assert clock.port == "PORT_A"
+    assert clock_submodule.get_total_clock_used() == 1
+
+def test_add_clock_duplicate_raises_exception():
+    mock_resources = Mock(spec=RsDeviceResources)
+    mock_resources.get_num_Clocks.return_value = 5
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    clock_submodule.add({"description": "Clock A", "port": "PORT_A", "enable": True})
+    
+    with pytest.raises(ClockDescriptionPortValidationException):
+        clock_submodule.add({"description": "Clock A", "port": "PORT_A", "enable": True})
+
+def test_add_clock_max_limit_reached():
+    mock_resources = Mock(spec=RsDeviceResources)
+    mock_resources.get_num_Clocks.return_value = 1
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    clock_submodule.add({"description": "Clock A", "port": "PORT_A", "enable": True})
+    
+    with pytest.raises(ClockMaxCountReachedException):
+        clock_submodule.add({"description": "Clock B", "port": "PORT_B", "enable": True})
+
+def test_remove_clock_success():
+    mock_resources = Mock(spec=RsDeviceResources)
+    mock_resources.get_num_Clocks.return_value = 5
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    clock_submodule.add({"description": "Clock A", "port": "PORT_A", "enable": True})
+    clock_submodule.add({"description": "Clock B", "port": "PORT_B", "enable": True})
+
+    removed_clock = clock_submodule.remove(0)
+
+    assert removed_clock.description == "Clock A"
+    assert len(clock_submodule.itemlist) == 1
+
+def test_remove_clock_not_found():
+    mock_resources = Mock(spec=RsDeviceResources)
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    with pytest.raises(ClockNotFoundException):
+        clock_submodule.remove(0)
+
+def test_get_all_messages():
+    mock_resources = Mock(spec=RsDeviceResources)
+    mock_resources.get_num_Clocks.return_value = 5
+    mock_resources.get_num_PLLs.return_value = 1
+
+    clock_submodule = Clock_SubModule(mock_resources)
+
+    clock_submodule.add({"description": "Clock A", "port": "PORT_A", "enable": False})
+    clock_submodule.add({"description": "Clock B", "port": "PORT_B", "enable": True})
+
+    mock_resources.get_VCC_CORE.return_value = 1.0
+    mock_resources.get_VCC_AUX.return_value = 1.2
+    mock_resources.get_PLL_INT.return_value = 1.0
+    mock_resources.get_PLL_AUX.return_value = 1.0
+    mock_resources.get_CLK_CAP.return_value = 0.5
+    mock_resources.get_CLK_INT_CAP.return_value = 0.3
+
+    mock_module = Mock()
+    mock_module.get_all.return_value = [
+        Mock(flip_flop=10, clock="PORT_A"),
+        Mock(flip_flop=5, clock="PORT_B")
+    ]
+    mock_resources.get_module.return_value = mock_module
+
+    mock_resources.get_clock_fanout = Mock(return_value=5)
+
+    clock_submodule.total_block_power = 50.0
+    clock_submodule.total_interconnect_power = 30.0
+    clock_submodule.total_pll_power = 10.0
+
+    mock_message = Mock()
+    mock_message.type = RsMessageType.INFO
+    clock_submodule.get_all_messages = Mock(return_value=[mock_message])
+
+    messages = clock_submodule.get_all_messages()
+
+    assert len(messages) > 0
+    assert messages[0].type == RsMessageType.INFO
