--- tests/test_rpe.py
@@ -1,9 +1,0 @@
-#
-#  Copyright (C) 2024 RapidSilicon
-#  Authorized use only
-#
-import pytest
-
-def test_rpe_1():
-    assert 1 == 1
-    # TODO: add backend test cases
